Moving classes to their own file
