+++ Sol.Domain/Models/BookClub.cs
@@ -1,0 +1,8 @@
+﻿namespace Sol.Domain.Models
+{
+    public record BookClub : Reading
+    {
+        public override HobbyType Type => HobbyType.BookClub;
+        public override string ToString() => "Book Club";
+    }
+}
